backend/src/requirements.txt
# Climaborough API - Backend Requirements

## Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
python-keycloak>=3.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Geospatial (if needed)
geoalchemy2>=0.14.0
shapely>=2.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
python-json-logger>=2.0.0

# Environment
python-dotenv>=1.0.0